Inference/requirements.txt
Flask
pandas
numpy
tensorflow-cpu
scikit-learn
joblib
requests
